backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
pytesseract==0.3.13
opencv-python==4.11.0.86
transformers==4.51.2
torch==2.6.0
pdf2image==1.17.0
pillow==11.2.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
web3==7.10.0
ipfshttpclient==0.7.0
python-magic==0.4.27
pydantic==2.11.3 